# docs/swagger.yaml
basePath: /api/v1
definitions:
  errorHandling.AppError:
    properties:
      code:
        type: integer
      key:
        type: string
      log:
        type: string
      message:
        type: string
    type: object
...
  models.Category:
    properties:
      Image:
        type: string
      Name:
        type: string
      Products:
        items:
          $ref: '#/definitions/models.Product'
        type: array
      ProductsStore:
        items:
          $ref: '#/definitions/models.ProductStore'
        type: array
      Slug:
        type: string
      created_at:
        type: integer
      id:
        $ref: '#/definitions/models.UID'
        type: object
      updated_at:
        type: integer
    type: object
...
  models.Image:
    properties:
      Url:
        type: string
      created_at:
        type: integer
      id:
        $ref: '#/definitions/models.UID'
        type: object
      updated_at:
        type: integer
    type: object
...
  models.Product:
    properties:
      ActiveElement:
        type: string
      Avatar:
        type: string
      Barcode:
        type: string
      BasePrice:
        type: number
      Category:
        items:
          $ref: '#/definitions/models.Category'
        type: array
      Code:
        type: string
      ConditionVoucher:
        type: string
      Content:
        type: string
      Cost:
        type: number
      Description:
        type: string
      Direction:
        type: string
      DoNotUse:
        type: string
      DrugInteractions:
        type: string
      GlobalManufacturerName:
        type: string
      HasPromote:
        type: boolean
      HasPromoteVoucher:
        type: boolean
      Images:
        items:
          $ref: '#/definitions/models.Image'
        type: array
      IndicationsOfTheDrug:
        type: string
      Manufacturer:
        type: string
      Name:
        type: string
      Overdose:
        type: string
      PackagingSize:
        type: string
      Percent:
        type: number
      RegistrationNo:
        type: string
      Status:
        type: string
      Storage:
        type: string
      Tags:
        items:
          $ref: '#/definitions/models.Tag'
        type: array
      Unit:
        type: string
      ValueVoucher:
        type: number
      Variants:
        items:
          $ref: '#/definitions/models.Variant'
        type: array
      Voucher:
        $ref: '#/definitions/models.Voucher'
        type: object
      VoucherId:
        $ref: '#/definitions/models.UID'
        type: object
      created_at:
        type: integer
      id:
        $ref: '#/definitions/models.UID'
        type: object
      updated_at:
        type: integer
    type: object
  models.ProductStore:
    properties:
      ActiveElement:
        type: string
      Avatar:
        type: string
      Barcode:
        type: string
      BasePrice:
        type: number
      Category:
        items:
          $ref: '#/definitions/models.Category'
        type: array
      Code:
        type: string
      Content:
        type: string
      Description:
        type: string
      Direction:
        type: string
      DoNotUse:
        type: string
      DrugInteractions:
        type: string
      GlobalManufacturerName:
        type: string
      Images:
        items:
          $ref: '#/definitions/models.Image'
        type: array
      IndicationsOfTheDrug:
        type: string
      Manufacturer:
        type: string
      Name:
        type: string
      Overdose:
        type: string
      PackagingSize:
        type: string
      RegistrationNo:
        type: string
      Status:
        type: string
      Storage:
        type: string
      Tags:
        items:
          $ref: '#/definitions/models.Tag'
        type: array
      Unit:
        type: string
      Variants:
        items:
          $ref: '#/definitions/models.Variant'
        type: array
      created_at:
        type: integer
      id:
        $ref: '#/definitions/models.UID'
        type: object
      updated_at:
        type: integer
    type: object
...
  models.Tag:
    properties:
      Name:
        type: string
      Slug:
        type: string
      created_at:
        type: integer
      id:
        $ref: '#/definitions/models.UID'
        type: object
      updated_at:
        type: integer
    type: object
  models.UID:
    type: object
...
  models.Variant:
    properties:
      Name:
        type: string
      created_at:
        type: integer
      id:
        $ref: '#/definitions/models.UID'
        type: object
      updated_at:
        type: integer
      variantValue:
        items:
          $ref: '#/definitions/models.VariantValue'
        type: array
    type: object
  models.VariantValue:
    properties:
      convertValue:
        type: number
      operator:
        type: string
      product:
        $ref: '#/definitions/models.Product'
        type: object
      productID:
        type: integer
      variant:
        $ref: '#/definitions/models.Variant'
        type: object
      variantID:
        type: integer
    type: object
  models.Voucher:
    properties:
      Deleted:
        type: boolean
      MaxValue:
        type: number
      Name:
        type: string
      Note:
        type: string
      Type:
        type: string
      Unit:
        type: string
      Value:
        type: number
      created_at:
        type: integer
      id:
        $ref: '#/definitions/models.UID'
        type: object
      updated_at:
        type: integer
    type: object
...
  requests.SearchTagRequest:
    properties:
      Name:
        example: name
        type: string
      Slug:
        example: slug
        type: string
      limit:
        example: 10
        type: integer
      offset:
        example: 0
        type: integer
      sort:
        $ref: '#/definitions/requests.SortOption'
        type: object
    type: object
  requests.SearchVariantRequest:
    properties:
      Name:
        example: name
        type: string
      limit:
        example: 10
        type: integer
      offset:
        example: 0
        type: integer
      sort:
        $ref: '#/definitions/requests.SortOption'
        type: object
    type: object
...
  requests.SortOption:
    properties:
      sort_direction:
        type: string
      sort_field:
        type: string
    type: object
  requests.TagRequest:
    properties:
      Name:
        example: name
        type: string
      Slug:
        example: slug
        type: string
    type: object
  requests.VariantRequest:
    properties:
      Name:
        example: name
        type: string
    type: object
...
  responses.Data:
    properties:
      code:
        type: integer
      key:
        type: string
      message:
        type: string
    type: object
...
  responses.TagSearch:
    properties:
      code:
        type: integer
      data:
        items:
          $ref: '#/definitions/models.Tag'
        type: array
      message:
        type: string
      total:
        type: integer
    type: object
...
  responses.VariantSearch:
    properties:
      code:
        type: integer
      data:
        items:
          $ref: '#/definitions/models.Variant'
        type: array
      message:
        type: string
      total:
        type: integer
    type: object
...
host: v1.api.medilane.vn
info:
  contact:
    name: medilane team
    url: https://www.medilane.vn/
  description: This is openapi for account api.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Medilane account api
  version: "1.0"
paths:
...
  /tag:
    post:
      consumes:
      - application/json
      description: Perform create Tag
      operationId: create-tag
      parameters:
      - description: Filter Tag
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/requests.TagRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/responses.Data'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorHandling.AppError'
      security:
      - BearerAuth: []
      summary: Create Tag in system
      tags:
      - Tag Management
  /tag/{id}:
    delete:
      consumes:
      - application/json
      description: Perform delete Tag
      operationId: delete-tag
      parameters:
      - description: id Tag
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.Data'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorHandling.AppError'
      security:
      - BearerAuth: []
      summary: Delete Tag in system
      tags:
      - Tag Management
    put:
      consumes:
      - application/json
      description: Perform edit Tag
      operationId: edit-tag
      parameters:
      - description: body Tag
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/requests.TagRequest'
      - description: id Tag
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.Data'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorHandling.AppError'
      security:
      - BearerAuth: []
      summary: Edit Tag in system
      tags:
      - Tag Management
  /tag/find:
    post:
      consumes:
      - application/json
      description: Perform search Tag
      operationId: search-tag
      parameters:
      - description: Filter Tag
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/requests.SearchTagRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.TagSearch'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorHandling.AppError'
      security:
      - BearerAuth: []
      summary: Search Tag in system
      tags:
      - Tag Management
  /variant:
    post:
      consumes:
      - application/json
      description: Perform create variant
      operationId: create-variant
      parameters:
      - description: Filter Variant
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/requests.VariantRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/responses.Data'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorHandling.AppError'
      security:
      - BearerAuth: []
      summary: Create variant in system
      tags:
      - Variant-Management
  /variant/{id}:
    delete:
      consumes:
      - application/json
      description: Perform delete variant
      operationId: delete-variant
      parameters:
      - description: id Variant
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.Data'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorHandling.AppError'
      security:
      - BearerAuth: []
      summary: Delete variant in system
      tags:
      - Variant-Management
    get:
      consumes:
      - application/json
      description: Perform edit variant
      operationId: get-variant
      parameters:
      - description: id Variant
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Variant'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorHandling.AppError'
      security:
      - BearerAuth: []
      summary: Edit variant in system
      tags:
      - Variant-Management
    put:
      consumes:
      - application/json
      description: Perform edit variant
      operationId: edit-variant
      parameters:
      - description: body Variant
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/requests.VariantRequest'
      - description: id Variant
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.Data'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorHandling.AppError'
      security:
      - BearerAuth: []
      summary: Edit variant in system
      tags:
      - Variant-Management
  /variant/find:
    post:
      consumes:
      - application/json
      description: Perform search variant
      operationId: search-variant
      parameters:
      - description: Filter Variant
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/requests.SearchVariantRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.VariantSearch'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorHandling.AppError'
      security:
      - BearerAuth: []
      summary: Search variant in system
      tags:
      - Variant-Management
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"